Create ReturnedEquipment
The input file used for the program. Contains the item type, a character designator for the item type, the hours contracted, the minutes over the contracted time, the customer number, and the payment type. Payment types are cash ($), credit (C), or EFT (E). If the payment is credit, then the card number is added. If the payment is EFT, then the routing number and account number are added.